+++ ecommerce/utils.py
@@ -183,9 +183,9 @@ def format_run_date(run_date):
 def deactivate_coupons(coupons, user_id=None):
     """
     Disables the provided coupons (enabled=False) and logs the deactivation with the given user, if specified.
+
     Args:
         coupons (List | QuerySet): Coupons to deactivate.
-        coupon_type (type): Coupon model for logging.
         user_id (int, optional): ID of the user performing the deactivation.
     Returns:
         set: Deactivated coupon codes and payment names.
@@ -197,6 +197,9 @@ def deactivate_coupons(coupons, user_id=None):
     content_type = ContentType.objects.get_for_model(Coupon)
 
     for coupon in coupons:
+        coupon.enabled = False
+        deactivated_codes_and_payment_names.add(coupon.coupon_code)
+        deactivated_codes_and_payment_names.add(coupon.payment.name)
         if user_id:
             log_entries.append(
                 LogEntry(
@@ -208,9 +211,6 @@ def deactivate_coupons(coupons, user_id=None):
                     change_message="Deactivated coupon",
                 )
             )
-        coupon.enabled = False
-        deactivated_codes_and_payment_names.add(coupon.coupon_code)
-        deactivated_codes_and_payment_names.add(coupon.payment.name)
 
     Coupon.objects.bulk_update(coupons, ["enabled"])
     LogEntry.objects.bulk_create(log_entries)
