+++ ecommerce/views_test.py
@@ -7,7 +7,9 @@ from urllib.parse import quote_plus, urljoin
 import factory
 import faker
 import pytest
+from django.contrib.auth.models import Permission
 from django.db.models import Count, Q
+from django.test import Client
 from django.urls import reverse
 from pytest_lazy_fixtures import lf as lazy
 from rest_framework import status
@@ -1214,6 +1216,49 @@ def test_create_coupon_permission(user_drf_client, promo_coupon_json):
     data = promo_coupon_json
     resp = user_drf_client.post(reverse("coupon_api"), type="json", data=data)
     assert resp.status_code == status.HTTP_403_FORBIDDEN
+
+
+@pytest.mark.parametrize(
+    (
+        "add_coupon",
+        "change_coupon",
+        "expected_admin_status",
+        "expected_coupons_status",
+        "expected_deactivate_status",
+    ),
+    [
+        (True, False, 200, 200, 403),
+        (False, True, 200, 403, 200),
+        (False, False, 403, 403, 403),
+        (True, True, 200, 200, 200),
+    ],
+)
+def test_ecommerce_restricted_view(  # noqa: PLR0913
+    user,
+    add_coupon,
+    change_coupon,
+    expected_admin_status,
+    expected_coupons_status,
+    expected_deactivate_status,
+):
+    """Test that the ecommerce restricted view is only accessible with the right permissions."""
+
+    user.user_permissions.clear()
+    if add_coupon:
+        user.user_permissions.add(Permission.objects.get(codename="add_coupon"))
+    if change_coupon:
+        user.user_permissions.add(Permission.objects.get(codename="change_coupon"))
+
+    client = Client()
+    client.force_login(user)
+
+    ecommerce_admin_url = reverse("ecommerce-admin")
+    add_coupons_url = ecommerce_admin_url + "coupons"
+    deactivate_coupons_url = ecommerce_admin_url + "deactivate-coupons"
+
+    assert client.get(ecommerce_admin_url).status_code == expected_admin_status
+    assert client.get(add_coupons_url).status_code == expected_coupons_status
+    assert client.get(deactivate_coupons_url).status_code == expected_deactivate_status
 
 
 def test_deactivate_coupons(mocker, admin_drf_client):
