+++ mitxpro/templates/index.html
@@ -5,8 +5,12 @@
 
 {% block content %}
 <div id="container"></div>
+{% render_bundle 'root' %}
+{% endblock %}
+
+{% block scripts %}
+{{ block.super }}
 <script>
   var USER_PERMISSIONS = {{user_permissions|safe}};
 </script>
-{% render_bundle 'root' %}
 {% endblock %}
